rights deletion

+++ Core.ApplicationServices/Contract/ItContractService.cs
@@ -24,7 +24,6 @@ namespace Core.ApplicationServices.Contract
 {
     public class ItContractService : IItContractService
     {
-        private readonly IGenericRepository<EconomyStream> _economyStreamRepository;
         private readonly IReferenceService _referenceService;
         private readonly ITransactionManager _transactionManager;
         private readonly IDomainEvents _domainEvents;
@@ -42,10 +41,10 @@ namespace Core.ApplicationServices.Contract
         private readonly IOptionsService<ItContract, PaymentFreqencyType> _paymentFrequencyOptionsService;
         private readonly IOptionsService<ItContract, OptionExtendType> _optionExtendOptionsService;
         private readonly IOptionsService<ItContract, TerminationDeadlineType> _terminationDeadlineOptionsService;
+        private readonly IEconomyStreamService _economyStreamService;
 
         public ItContractService(
             IItContractRepository repository,
-            IGenericRepository<EconomyStream> economyStreamRepository,
             IReferenceService referenceService,
             ITransactionManager transactionManager,
             IDomainEvents domainEvents,
@@ -61,10 +60,10 @@ namespace Core.ApplicationServices.Contract
             IOptionsService<ItContract, PaymentModelType> paymentModelOptionsService,
             IOptionsService<ItContract, PaymentFreqencyType> paymentFrequencyOptionsService,
             IOptionsService<ItContract, OptionExtendType> optionExtendOptionsService,
-            IOptionsService<ItContract, TerminationDeadlineType> terminationDeadlineOptionsService)
+            IOptionsService<ItContract, TerminationDeadlineType> terminationDeadlineOptionsService, 
+            IEconomyStreamService economyStreamService)
         {
             _repository = repository;
-            _economyStreamRepository = economyStreamRepository;
             _referenceService = referenceService;
             _transactionManager = transactionManager;
             _domainEvents = domainEvents;
@@ -81,6 +80,7 @@ namespace Core.ApplicationServices.Contract
             _paymentFrequencyOptionsService = paymentFrequencyOptionsService;
             _optionExtendOptionsService = optionExtendOptionsService;
             _terminationDeadlineOptionsService = terminationDeadlineOptionsService;
+            _economyStreamService = economyStreamService;
         }
 
         public Result<ItContract, OperationError> Create(int organizationId, string name)
@@ -134,36 +134,38 @@ namespace Core.ApplicationServices.Contract
             {
                 return OperationFailure.Forbidden;
             }
-            using (var transaction = _transactionManager.Begin())
+
+            using var transaction = _transactionManager.Begin();
+            try
             {
-                try
+                //Delete the economy streams to prevent them from being orphaned
+                var economyStreams = _economyStreamService.GetEconomyStreams(contract);
+                var deleteEconomyStreamsResult = _economyStreamService.DeleteRange(economyStreams);
+                if (deleteEconomyStreamsResult.HasValue)
                 {
-                    //Delete the economy streams to prevent them from being orphaned
-                    foreach (var economyStream in GetEconomyStreams(contract))
-                    {
-                        DeleteEconomyStream(economyStream);
-                    }
-                    _economyStreamRepository.Save();
-
-                    //Delete the contract
-                    var deleteByContractId = _referenceService.DeleteByContractId(id);
-                    if (deleteByContractId.Failed)
-                    {
-                        transaction.Rollback();
-                        return deleteByContractId.Error;
-                    }
-                    _domainEvents.Raise(new EntityBeingDeletedEvent<ItContract>(contract));
-                    _repository.DeleteContract(contract);
-
-                    transaction.Commit();
-                }
-                catch (Exception e)
-                {
-                    _logger.Error(e, $"Failed to delete it contract with id: {contract.Id}");
                     transaction.Rollback();
-                    return OperationFailure.UnknownError;
+                    return deleteEconomyStreamsResult.Value.FailureType;
                 }
+
+                //Delete the contract
+                var deleteByContractId = _referenceService.DeleteByContractId(id);
+                if (deleteByContractId.Failed)
+                {
+                    transaction.Rollback();
+                    return deleteByContractId.Error;
+                }
+                _domainEvents.Raise(new EntityBeingDeletedEvent<ItContract>(contract));
+                _repository.DeleteContract(contract);
+
+                transaction.Commit();
             }
+            catch (Exception e)
+            {
+                _logger.Error(e, $"Failed to delete it contract with id: {contract.Id}");
+                transaction.Rollback();
+                return OperationFailure.UnknownError;
+            }
+
             return contract;
 
         }
@@ -286,6 +288,16 @@ namespace Core.ApplicationServices.Contract
                 );
         }
 
+        public Result<ItContract, OperationError> RemoveContractResponsibleUnit(int contractId)
+        {
+            return Modify(contractId, contract =>
+            {
+                contract.ResponsibleOrganizationUnit = null;
+                contract.ResponsibleOrganizationUnitId = null;
+                return Result<ItContract, OperationError>.Success(contract);
+            });
+        }
+
         private Result<ContractOptions, OperationError> WithOrganizationReadAccess(int organizationId, Func<Result<ContractOptions, OperationError>> authorizedAction)
         {
             var readAccessLevel = _authorizationContext.GetOrganizationReadAccessLevel(organizationId);
@@ -303,19 +315,6 @@ namespace Core.ApplicationServices.Contract
         private Result<ItContract, OperationError> WithReadAccess(ItContract contract)
         {
             return _authorizationContext.AllowReads(contract) ? Result<ItContract, OperationError>.Success(contract) : new OperationError(OperationFailure.Forbidden);
-        }
-
-        private static IEnumerable<EconomyStream> GetEconomyStreams(ItContract contract)
-        {
-            return contract
-                .ExternEconomyStreams
-                .ToList()
-                .Concat(contract.InternEconomyStreams.ToList());
-        }
-
-        private void DeleteEconomyStream(EconomyStream economyStream)
-        {
-            _economyStreamRepository.DeleteWithReferencePreload(economyStream);
         }
 
         private Result<TSuccess, OperationError> Modify<TSuccess>(int id, Func<ItContract, Result<TSuccess, OperationError>> mutation)
